requirements.txt
absl-py==2.1.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
easyocr==1.7.1
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.48.1
fqdn==1.5.1
fsspec==2024.6.0
gast==0.5.4
glib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.23.2
idna==3.4
imageio==2.34.1
importlib_metadata==7.1.0
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kaggle==1.6.6
keras==3.3.3
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mplcursors==0.5.3
mpmath==1.3.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
onnx==1.16.1
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
opendatasets==0.1.22
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
pango==0.0.1
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf=3.20
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycparser==2.21
pydyf==0.10.0
Pygments==2.16.1
pyparsing==3.1.1
pypdfium2==4.30.0
pyphen==0.15.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-doctr==0.8.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
rapidfuzz==3.9.3
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==68.2.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-macos==2.16.1
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
tifffile==2024.5.22
tinycss2==1.3.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.1
tzdata==2023.3
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.9
weasyprint==62.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.3
wheel==0.41.3
widgetsnbextension==4.0.9
wrapt==1.16.0
zipp==3.19.1
zopfli==0.2.3
